gemini/requirements.txt
google-cloud-aiplatform
google-cloud-texttospeech
google-genai